Add metadata(description, edition, product type, release date, publisher name and rating information)  extraction and DB refactoring

+++ src/main/java/io/github/Mishaa105/price_tracker/db/repository/OfferRepository.java
@@ -3,4 +3,4 @@ package io.github.Mishaa105.price_tracker.db.repository;
 import io.github.Mishaa105.price_tracker.db.entity.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OfferRepository extends JpaRepository<Offer, Long> {}
+public interface OfferRepository extends JpaRepository<Offer, Integer> {}
